Ensure flickable velocity is updated when under heavy load.
If the Flickable is forced to move very quickly due to highlight motion
the smoothed velocity animation may not have an opportunity to run
before being reset, resulting in the velocity not being updated.  In this
case just set the velocity directly since the motion is driven by an
animation anyway.

Change-Id: I1e5e9396f4c1cd12e95c6ac47941dad25e4aa897
Reviewed-by: Andrew den Exter <andrew.den-exter@nokia.com>