Document function limitation of var properties
Functions cannot be assigned to var properties due to the fact that
such an assignment signifies a binding assignment.  This limitation
needs to be documented.  One workaround is to assign an array which
contains a function element to a var property, and this commit also
adds a unit test to ensure this works.

Change-Id: I02363b88233282106ac6d26f14df1988155057b9
Reviewed-by: Martin Jones <martin.jones@nokia.com>