Fix loading imports from Qt resource.
Loading an import from a Qt resource fails because the url of the module
was erroneously being converted to a file:// url.  This results in a url
that looks like "file://:/MyModule".  Fixed by adding a check to see if
the module path is a resource path (starts with :) and if so constructs
a qrc:// url instead.

Change-Id: I41a24a93d085baaf68d58c4e991d6ba2aff5e7ff
Reviewed-by: Martin Jones <martin.jones@nokia.com>