Prevent consecutive images from climbing up a text node.
Vertically position all nodes at the baseline minus their ascent.
Previously the bounding rect of a glyph run was postioned on the base
line and  images at the baseline minus the ascent.  This meant that
calculated position of an image was correct if it followed text but
if it followed another image it would be positioned above it.

Change-Id: I03f5e0b48f132f010d16c49620fa9463873f9492
Reviewed-by: Yann Bodson <yann.bodson@nokia.com>