Fix SpringAnimation and SmoothedAnimation issues
In the new animation framework, we now create a new parallel animation wrapper
group from each call to transition (in smoothedanimation and springanimation),
and putting the old qsmoothedanimations/qspringanimations into this new wrapper
forces them to stop and then start rather than continue.

This will cause problems if the transition() function is called too frequently
(for example, called more than twice between 16ms).

To fix this issue, introduce a currentDelta() function to QUnifyTimer2 class
which can be used for the animations which need to be restarted.

Specially for the QSpringAnimation, it can always use the currentDelta() rather
than calculating the elapsed time from the last time and the current time.

Change-Id: I46ade84165ce1d42e70e048021105ae072734e29
Reviewed-by: Michael Brasser <michael.brasser@nokia.com>