Add accessibility to qml
This consists of two parts:
An attached property for QML items and a plugin for the accessibility framework.

The attached property simply takes care of some properties (name, role)
that are needed in order to expose semantics of the application to
assistive tools.

The plugin exposes the hierarchy of QML items to the
accessibility framework.

Change-Id: I32f5603d0d9549b01b3645b205b710b9801762f7
Reviewed-by: Jan-Arve Sæther <jan-arve.saether@nokia.com>