Fix tests broken by a change in the signals emitted by AbstractItemModel
QAbstractItemModel used to emit both layoutChanged() and rowsMoved()
when items were moved.  The VisualDataModel test expected both signals
but was only interested in the move, and the grid view test had a bug
in its test model that didn't was hidden by the change signal being
emitted and causing bindings to be reevaluated.

Change-Id: Id5d1b5768717f4a82f6eb4710ff3b4429529193e
Reviewed-by: Alan Alpert <alan.alpert@nokia.com>