Debugger: Fix livelock on exit (Windows)
We used to close the debugger thread when the last service was killed: This happened
in the unloading of static global variables. Apparently this is too late on Windows,
since the OS thread associated was already terminated.

Instead, we now kill the instance when the QCoreApplication event loop is exiting.

Change-Id: I12a46ab9e7ac64561c94c0cd0d88b78fbaf8554c
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@nokia.com>
Reviewed-by: Aurindam Jana <aurindam.jana@nokia.com>