VisualDataModel performance improvements.
Avoid a linear scan of all cache items and associated accesses by
getting the cache item from an objects  vdm attached object instead.

Make the model context property a property of the context object instead
of a separate property on the context object.

Parent the vdm attached object to the delegate object with
QDeclarative_setParent_noEvent instead of passing it in the constructor.

Change-Id: Ib77c5cdb963f3dfe8f2bdef039e010a6bb30140f
Reviewed-by: Martin Jones <martin.jones@nokia.com>