Handle exceptions while compiling v8 bindings
Previously, no exception handling existed, which could cause a crash
if an invalid v8 binding expression was generated.  Such invalid
bindings should usually be rewritten into valid form by the bindings
rewriter, but in some cases it is too costly to do so, so we need
to handle exceptions.

Task-number: QTBUG-24064
Task-number: QTBUG-23387
Change-Id: I7da12a936780a561c9e9cad3a4a7b62c06d6973e
Reviewed-by: Alan Alpert <alan.alpert@nokia.com>