Make QAccessibleQuickView::childAt() work properly with overlapping items
The previous code did not consider items that were overlapped due to
having different z coordinates.

The approach used is now the same as found in
QQuickCanvas::mousePressEvent().

Strictly speaking, this is a violation of childAt (since it will
disregard the implementation of childAt of all the descendants along
the path down to the item actually returned.)

However, I don't see any good reason for that the implementation for
childAt() would be different than how mousePressEvent behaves.
It should also perform better than any other solution I managed to
think of.

Change-Id: I2d3fa2282437c7b5533c6149c62fc456ccf2ccfa
Reviewed-by: Frederik Gladhorn <frederik.gladhorn@nokia.com>