Fix moving mixed visible and non-visible View items.
A removal for a move that included both cached and non-cached was
only evaluating the first part of the removal and skipping the
remainder.

Don't skip all subsequent removals when two ranges are merged, instead
reset the current indexes to the start of the merged range and continue
on as normal.

Task-number: QTBUG-23432
Change-Id: Ieda49e2f68cd0bbbd31ba1e7dfd5af7c3e350d78
Reviewed-by: Bea Lam <bea.lam@nokia.com>