Per-frame Sprites patch one
To allow for sprites to be advanced by the rendering framerate, two
minor redesigns were needed.

A) Sprite texture location is now calculated on the CPU and passed to
the GPU per frame.

B) Stochastic State engine now supports states that do not advance on a
timer, and states can be advanced manually.

This patch implements B and A for ImageParticle. A for SpriteImage will
be done in a separate patch.

Task-number: QTBUG-22236
Change-Id: If1c54a6a03fa48b95bb1e672283292859656457b
Reviewed-by: Alan Alpert <alan.alpert@nokia.com>