Fix preedit text visibility for TextEdit
For an empty TextEdit with preedit text, we need to add the text
separately since the blockIterator will not go through that case.
Also fixed preedit text appended at the end of any commited text.

Move the code for adding text into SelectionEngine so that we can
reuse it for adding preedit text.

Task-number: QTBUG-22647
Change-Id: Id4fe04099b16949ff5de0747881c6dc96ef0673f
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@nokia.com>