Test more of QQuickImageParticle
Now the color variance properties are tested, as is the debug code.
Just running the debugging code is valid testing, as it contains
pointer derefernces that could cause runtime errors, and this codepath
is not always enabled at runtime.

Additionally, debug mode control has been consolidated in the
ParticleSystem. However, this was not necessary for this test addition.

Change-Id: Ie56465145461486456462154dfafe546fedabcba
Reviewed-by: Martin Jones <martin.jones@nokia.com>