Remove QJSValue::instanceOf() function
Rationale: This is a remnant from QtScript. There is no good reason
for providing this type of low-level "prototype inheritance chain"
checks in this high-level QJSValue class. If you want to check if an
object is of the "right type", you can check if it has the
properties you require using property().

Task-number: QTBUG-23604
Change-Id: I3a274212cc57c38228fab73423af481e1b95d8a5
Reviewed-by: Simon Hausmann <simon.hausmann@nokia.com>
Reviewed-by: Jędrzej Nowacki <jedrzej.nowacki@nokia.com>