Remove QJSEngine::newVariant() function
Rationale: This is a remnant from QtScript. There is no reason one
should be able to explicitly wrap an arbitrary C++ type in a
QVariant. toScriptValue() will automatically wrap the type in a
QVariant if it's a type that's not specifically handled by the
meta-type system, or if it can't usefully/losslessly be converted to
a native JavaScript type.

Task-number: QTBUG-23604
Change-Id: I5790cd92b9ffd5fcfbee1cd73a8eaa16fc74846f
Reviewed-by: Jędrzej Nowacki <jedrzej.nowacki@nokia.com>