Remove QJSValue::propertyFlags() function
Rationale: The API is incomplete. There is no way to set/update the
flags; adding a third (optional) argument to setProperty() is ugly.

If necessary, a QJSPropertyDescriptor API should be added, to enable
complete setting/querying of properties.

The flags were moved to QJSValuePrivate, and the propertyFlags()
implementation kept, because there is one place the read-only flag
is checked internally (in qv8typewrapper).

Task-number: QTBUG-23604
Change-Id: I089dce28dbda428662aac7c8d1536987fe6d855c
Reviewed-by: Jędrzej Nowacki <jedrzej.nowacki@nokia.com>