Remove QJSEngine::nullValue() function
Rationale: It's strange to have a null value factory function.
There should just be one way of constructing null values: By
passing NullValue to the QJSValue constructor.

The nullValue() function created a value that was bound to the
engine; the QJSValue constructor does not. In order to ensure that
we're testing the same behavior as before, I've replaced nullValue()
calls by evaluate("null") in the autotests.

Task-number: QTBUG-23604
Change-Id: Ie40b61fa64e070b90c6245fd21554963073c5f80
Reviewed-by: Jędrzej Nowacki <jedrzej.nowacki@nokia.com>