Remove QJSEngine::undefinedValue() function
Rationale: It's strange to have an undefined value factory function.
There should just be one way of constructing undefined values: By
passing UndefinedValue to the QJSValue constructor.

The undefinedValue() function created a value that was bound to the
engine; the QJSValue constructor does not. In order to ensure that
we're testing the same behavior as before, I've replaced
undefinedValue() calls by toScriptValue(QVariant()) in the autotests.

Task-number: QTBUG-23604
Change-Id: Ie9295b7af49c853b5a4a3445da48c325fef8d6b1
Reviewed-by: Jędrzej Nowacki <jedrzej.nowacki@nokia.com>
Reviewed-by: Simon Hausmann <simon.hausmann@nokia.com>