Accessibility: childAt_helper descends too far.
Recursive hit-testing using childAt_helper must stop
when it hits a leaf accessibility item. This leaf item
might not be a leaf item in Qt Quick - a button contains
text and mouse area items not visible to the accessibility
system for example.

Get the accessible interface and check if it has
children before recursing to the child items.

Change-Id: I2cc286cde4bce6b4ca340e8a1819edc5f84006a5
Reviewed-by: Jan-Arve Sæther <jan-arve.saether@nokia.com>
Reviewed-by: Frederik Gladhorn <frederik.gladhorn@nokia.com>