tst_qdeclarativedebugjs: Use new QJSValue::call() overload
The old QJSValue::call() function that takes a this-object as the
first argument is going away. The new overload simply takes a list
of arguments, so that the "dummy" this-object doesn't have to be
explicitly passed.

The new function callWithInstance() can be used to call a function
where you want to use a specific this-object.

Change-Id: I2d9bd79f8e5d00a2b89568171ccc5ffcee52686c
Reviewed-by: Jędrzej Nowacki <jedrzej.nowacki@nokia.com>
Reviewed-by: Kai Koehne <kai.koehne@nokia.com>