Remove deprecated QtDeclarative (now QtQuick) headers
The QtQuick module has been around for a while now, and the need to
port to it was duly announced, coupled with a qmake warning (which
has now been removed).

Users who haven't ported their includes over will have to do so as
of this commit.

The fixqt4headers script in qtbase can be used to automatically
update the stale include statements in existing projects.

Change-Id: I9745723246f65ef726e51b8551d18b378085689d
Reviewed-by: Lars Knoll <lars.knoll@nokia.com>