QQuickCanvas::event should return true if the touch event was accepted
If QQuickCanvas::event delivers a touch event and it is accepted
the control ends up in QWindow::event which invalidates the event.
These touch events end up as if they were unhadled which causes Qt to
automatically synthesize mouse events even for accepted touch events.

Add a unit test for testing this behavior.

Change-Id: I83d4aeafee1ea7ec5d219e4b45aae699188717c3
Reviewed-by: Zeno Albisser <zeno.albisser@nokia.com>
Reviewed-by: Samuel Rødal <samuel.rodal@nokia.com>