Accept default targets for transitions
QDeclarativeAbstractAnimation::transition(),
QDeclarativeTransition::prepare() and QDeclarativeTransitionManager
now accept a QObject* to specify a default target for a transition.
For view transitions, this will allow the view to pass a view item
as the default target for a view transition so that the Transition
does not have to explicitly set the target for every animation.

Change-Id: I2c8d5535283335e139d98ef56f399f338627fb3e
Reviewed-by: Michael Brasser <michael.brasser@nokia.com>