Fix acceptance of unhandled mouse events.
Mouse events not handled by the canvas need to be ignored, in order for
features like touch synthetization to work. In case of the initially
delivered mouse press event, the canvas did not ignore the event if it
was not handled. This patch corrects that and adds an auto-test to verify
the new behaviour.

Change-Id: I8499701f5a161ce1e90a70303aa7ca4ccf2f0b6f
Reviewed-by: Andras Becsi <andras.becsi@nokia.com>
Reviewed-by: Samuel Rødal <samuel.rodal@nokia.com>