Use use touch events to generate mouse events in Quick canvas.
Allowing the canvas to deal with mouse event emulation provides
a couple of advantages: 1. touches outside a reactive area can
be ignored, allowing a touch within a reactive area to become
the "mouse touch". 2. we can extend the mouse events with
data available in touch events for use within the quick elements;
specifically velocity in this case.

Change-Id: Id9df397f1a3a03dbda2ef5dcf6189e7eb3377356
Reviewed-by: Michael Brasser <michael.brasser@nokia.com>