Add linkColor property to Text.
Allows the color of links in text to be changed from the default blue.

This currently only works with StyledText and the distance field
rendererer.  It could be made to work with RichText overwriting the
specified foreground color in all instances or by not setting a default
color in the html parser.  The former would prevent the color being
set with CSS or some future means for altering text formats.  The latter
would break rendering with QPainter.

Task-number: QTBUG-23048
Change-Id: I98df215cabe8a089f648fd4a6206622b4318fb8f
Reviewed-by: Martin Jones <martin.jones@nokia.com>