qmltest: Count passes, fails and skips consistently.
This commit changes qmltest to count one pass per data row, rather than
a single pass for an entire data-driven test function.  This makes
counting of passes consistent with counting of fails and skips which
were already counted once per data row.  This change also means that the
plain-text, xml and light-xml output formats will now show one result
per data row, so that every data row executed will be shown in the test
output with either a pass, fail or skip result.  Previously data rows
that passed were not shown in the output, preventing analysis of total
number of tests runs, pass-rates, and various other metrics.

This commit corresponds to (and depends on) a commit in the qtbase
module which changes the behaviour of qtestlib in the same way.

Task-number: QTBUG-21848
Task-number: QTBUG-22124
Change-Id: I3c1f0e68bdff4087b9ccfc80a3f96f4541335b6f
Reviewed-by: Rohan McGovern <rohan.mcgovern@nokia.com>
Reviewed-by: Yunqiao Yin <charles.yin@nokia.com>