Create a forwarding declarative module.
The declarative module has been replaced by the qml module.  To allow
existing projects which depend on the old declarative module to build,
add a new declarative module which forwards dependencies on to the
new qml module.

This module contains only redefinitions of old symbol names to their
new replacements, and creates forwarding headers that include the
replacement headers provided by the qml module.

This is a temporary change to prevent breaking existing projects
that depend on the declarative module. After clients have had an
opportunity to update their code to the use the new interfaces,
it can be removed.

Task-number: QTBUG-23737
Change-Id: I3cec5ceec4f9725c258e3672360a05ea87d3ca2f
Reviewed-by: Martin Jones <martin.jones@nokia.com>