Update item position in parent change atomically.
Changing x and y individually generates two geometry changed events,
the first of which has an invalid position as the x coordinate is
relative to the new parent and the y relative to the old parent.

This in turn causes the Drag item to send move events with incorrect
positions.

Task-number: QTBUG-24534
Change-Id: If2636a968acc0fffce21d1a7e51510426ace38a0
Reviewed-by: Michael Brasser <michael.brasser@nokia.com>