QQuickView with the default constructor doesn't show any window decorations.
The default parameter takes a 0 which is propagated to the window platform
which sets it to 0. QWidget have a method called adjustFlags which takes
care of adding QWindow when necessary. There is no such check in QWindow
or the platform plugin for QPA. The default flag of QWindow is Qt::Window
but in that case it is overiden as QQuickView constructor calls
setWindowFlags(0). Qt::Window as a default parameter is a solution.

Change-Id: Iccbebc5b7e4f094c7fc6dc61b37d9f9622be0237
Reviewed-by: Samuel Rødal <samuel.rodal@nokia.com>
Reviewed-by: Gunnar Sletta <gunnar.sletta@nokia.com>